+++ embed-pdf-viewer.php
@@ -14,7 +14,7 @@
  * Description:       Embed a PDF from the Media Library or elsewhere via oEmbed or as a block into an `object` tag or Google Doc Viewer as fallback.
  * Author:            Andy Fragen
  * Author URI:        https://github.com/afragen
- * Version:           2.0.5.2
+ * Version:           2.0.5.3
  * License:           GPLv2+
  * Domain Path:       /languages
  * Text Domain:       embed-pdf-viewer
@@ -166,9 +166,10 @@ class Embed_PDF_Viewer {
 		}
 
 		$default = [
-			'height' => 500,
-			'width'  => 800,
-			'title'  => $post->post_title,
+			'height'      => 500,
+			'width'       => 800,
+			'title'       => $post->post_title,
+			'description' => $post->post_content,
 		];
 
 		/*
@@ -191,24 +192,26 @@ class Embed_PDF_Viewer {
 		$atts = apply_filters( 'embed_pdf_viewer_pdf_attributes', $atts );
 
 		// Fix title or create from filename.
-		$atts['title'] = empty( $atts['title'] )
+		$atts['title']       = empty( $atts['title'] )
 			? ucwords( preg_replace( '/(-|_)/', ' ', $post->post_name ) )
 			: ucwords( preg_replace( '/(-|_)/', ' ', $atts['title'] ) );
+		$atts['description'] = empty( $atts['description'] ) ? $atts['title'] : $atts['description'];
 
 		$iframe_fallback  = '<iframe class="embed-pdf-viewer" src="https://docs.google.com/viewer?url=' . rawurlencode( $post->guid );
 		$iframe_fallback .= '&amp;embedded=true" frameborder="0" ';
 		$iframe_fallback .= 'style="height:' . $atts['height'] . 'px;width:' . $atts['width'] . 'px;" ';
-		$iframe_fallback .= 'title="' . $atts['title'] . '"></iframe>' . "\n";
+		$iframe_fallback .= 'title="' . $atts['description'] . '"></iframe>' . "\n";
 
 		$object  = '<object class="embed-pdf-viewer" data="' . $post->guid;
 		$object .= '#scrollbar=1&toolbar=1"';
 		$object .= 'type="application/pdf" ';
-		$object .= 'height=' . $atts['height'] . ' width=' . $atts['width'] . ' > ';
+		$object .= 'height=' . $atts['height'] . ' width=' . $atts['width'] . ' ';
+		$object .= 'title="' . $atts['description'] . '"> ';
 		$object .= '</object>';
 
 		$embed  = '<figure>';
 		$embed .= $object . $iframe_fallback;
-		$embed .= '<p><a href="' . $post->guid . '">' . $atts['title'] . '</a></p>';
+		$embed .= '<p><a href="' . $post->guid . '" title="' . $atts['description'] . '">' . $atts['title'] . '</a></p>';
 		$embed .= '</figure>';
 
 		return $embed;