+++ CHANGES.md
@@ -1,5 +1,6 @@
 #### [unreleased]
 * update block build process
+* add title attributes for a11y
 
 #### 2.0.5 / 2020-03-09
 * minor updates to plugin structure on GitHub